# backend/src/user/dto/user.dto.ts
import { ApiProperty } from "@nestjs/swagger";

export class UserDto{
    @ApiProperty()
    id: number;

    @ApiProperty()
    email: string;

    @ApiProperty()
    name: string;

    @ApiProperty({ nullable: true, type: String })
    imageUrl?: string | null;
}